requirements.txt
fastapi
uvicorn[standard]
boto3
pandas
jupyter
scikit-learn
mlflow
kagglehub
seaborn
xgboost
python-dotenv
psycopg2
scikit-learn
SQLAlchemy
streamlit
plotly